refactor(ui): relativetime 컴포넌트를 공통 ui로 이동
재사용성을 위해 공용 컴포넌트로 이동해줬습니다
